move term to trait module + add IsPolynomial trait
